refactor: 게시글 작성 수정

+++ src/main/java/com/mini/hanghae99miniproject/post/service/PostService.java
@@ -1,9 +1,9 @@
 package com.mini.hanghae99miniproject.post.service;
 
+import com.mini.hanghae99miniproject.comment.entity.Comment;
 import com.mini.hanghae99miniproject.comment.repository.CommentRepository;
 import com.mini.hanghae99miniproject.member.entity.Member;
 import com.mini.hanghae99miniproject.post.dto.RequestPostDto;
-import com.mini.hanghae99miniproject.post.dto.ResponseAllPostDto;
 import com.mini.hanghae99miniproject.post.dto.ResponsePostDto;
 import com.mini.hanghae99miniproject.post.entity.Post;
 import com.mini.hanghae99miniproject.post.mapper.PostMapper;
@@ -16,6 +16,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static com.mini.hanghae99miniproject.common.exception.ExceptionMessage.NO_EXIST_POSTING_ERROR_MSG;
+import static com.mini.hanghae99miniproject.common.exception.ExceptionMessage.USER_NOT_MATCH_ERROR_MSG;
 
 
 @Service
@@ -33,17 +34,17 @@ public class PostService {
         Post post = postMapper.toEntity(requestPostDto, member);
         postRepository.save(post);
 
-        return postMapper.postToResponsePostDto(post);
+        return postMapper.postToResponsePostDtoALL(post);
     }
 
     //게시글 전체 조회
     @Transactional(readOnly = true)
-    public List<ResponseAllPostDto> findAllPost() {
+    public List<ResponsePostDto> findAllPost() {
         //DB에 저장되어있는 게시글 전부 가져오기
         List<Post> postList = postRepository.findAll();
-        List<ResponseAllPostDto> result = new ArrayList<>();
+        List<ResponsePostDto> result = new ArrayList<>();
         for (Post post : postList) {
-            result.add(new ResponseAllPostDto(post));
+            result.add(postMapper.postToResponsePostDtoALL(post));
         }
         return result;
     }
@@ -56,27 +57,45 @@ public class PostService {
                 () -> new IllegalArgumentException(NO_EXIST_POSTING_ERROR_MSG.getMsg())
         );
 
-        return new ResponsePostDto(post);
+        //선택한 게시글에서 작성한 댓글만 조회
+        List<Comment> comments = commentRepository.findAllByPostOrderByCreatedAtDesc(post);
+        List<ResponseComment> commentList = new ArrayList<>();
+        for (Comment comment : comments) {
+            commentList.add(new ResponseComment(comment));
+        }
+
+        return postMapper.postToResponsePostDto(post, commentList);
     }
 
     //게시글 수정
+    @Transactional
     public ResponsePostDto updatePost(Long id, RequestPostDto requestPostDto, Member member) {
         //게시글이 있는지 없는지 조회 없으면 예외처리
         Post post = postRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException(NO_EXIST_POSTING_ERROR_MSG.getMsg())
         );
 
-        post.update(requestPostDto.getTitle(), requestPostDto.getContent());
-        postRepository.save(post);
+        // 게시글 작성자와 멤버 정보가 일치하는지 확인
+        if(!post.getMember().getId().equals(member.getId())) {
+            throw new IllegalArgumentException(USER_NOT_MATCH_ERROR_MSG.getMsg());
+        }
 
-        return new ResponsePostDto(post);
+        post.update(requestPostDto.getTitle(), requestPostDto.getContent());
+
+        return postMapper.postToResponsePostDtoALL(post);
     }
 
     //게시글 삭제
+    @Transactional
     public void deletePost(Long id, Member member) {
         Post post = postRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException(NO_EXIST_POSTING_ERROR_MSG.getMsg())
         );
+
+        // 게시글 작성자와 멤버 정보가 일치하는지 확인
+        if(!post.getMember().getId().equals(member.getId())) {
+            throw new IllegalArgumentException(USER_NOT_MATCH_ERROR_MSG.getMsg());
+        }
 
         // 해당 게시글에 관련된 댓글들 삭제
         commentRepository.deleteCommentByPost(post);
